fix: 건물 메타데에터 종류 캐싱으로 수정
- 쿼리로 level이 1 yield만 join해 가져오도록 수정
