+++ src/main/java/store/sonyk9919/api/domain/building/service/BuildingMetadataQueryService.java
@@ -1,28 +1,19 @@
 package store.sonyk9919.api.domain.building.service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import store.sonyk9919.api.domain.building.dto.BuildingCatalogDto;
-import store.sonyk9919.api.domain.building.entity.BuildingMetadata;
-import store.sonyk9919.api.domain.building.repository.BuildingMetadataRepository;
 
 @Service
 @RequiredArgsConstructor
 public class BuildingMetadataQueryService {
 
-    private final BuildingMetadataRepository metadataRepository;
+    private final BuildingMetadataCache metadataCache;
 
     @Transactional(readOnly = true)
     public List<BuildingCatalogDto> getBuildingCatalog() {
-        return metadataRepository.findAll().stream()
-                .map(this::mapToCatalog)
-                .collect(Collectors.toList());
-    }
-
-    private BuildingCatalogDto mapToCatalog(BuildingMetadata metadata) {
-        return BuildingCatalogDto.of(metadata, metadata.getYieldForLevel(1));
+        return metadataCache.get();
     }
 }
